refactor: migrate page 404 to app router
